[Refactor] Split ImportExportWizard into per-step files
Break the monolithic 1341-line index.tsx into:

- steps/StepPickFile.tsx
- steps/StepSummary.tsx
- steps/StepGlobalSettings.tsx
- steps/StepPerGame.tsx
- steps/StepCredentials.tsx
- steps/StepLibrarySystem.tsx
- steps/StepDone.tsx
- steps/index.ts (barrel)
- shared.tsx — PathActionRow, StatusPill, runnerLabel(),
  stageFriendlyLabel() and the PathChoice / PathAction types used
  across multiple steps

index.tsx shrinks to 379 lines and now owns only the wizard
orchestration (state, navigation, the apply call) plus the dialog
chrome. No behavioural change; codecheck, lint, full test suite and
electron-vite build all clean.
