+++ src/frontend/components/UI/ImportExportWizard/steps/StepGlobalSettings.tsx
@@ -1,0 +1,91 @@
+import { useTranslation } from 'react-i18next'
+import WarningAmberIcon from '@mui/icons-material/WarningAmber'
+
+import ToggleSwitch from 'frontend/components/UI/ToggleSwitch'
+
+import type { HeroicBackupValidationReport } from 'common/types/importExport'
+
+interface Props {
+  validation: HeroicBackupValidationReport
+  includeGlobal: boolean
+  setIncludeGlobal: (v: boolean) => void
+  overwriteGlobal: boolean
+  setOverwriteGlobal: (v: boolean) => void
+}
+
+export default function StepGlobalSettings({
+  validation,
+  includeGlobal,
+  setIncludeGlobal,
+  overwriteGlobal,
+  setOverwriteGlobal
+}: Props) {
+  const { t } = useTranslation()
+  const { manifest } = validation
+  const hasGlobal = manifest.stages.includes('globalSettings')
+  return (
+    <section>
+      <h3 className="ImportExportWizard__heading">
+        {t('import-export.step3.title', 'Global settings, themes and fixes')}
+      </h3>
+      {!hasGlobal ? (
+        <p className="ImportExportWizard__hint">
+          {t(
+            'import-export.step3.none',
+            'This backup does not contain global settings.'
+          )}
+        </p>
+      ) : (
+        <>
+          <ToggleSwitch
+            htmlId="ie-include-global"
+            title={t(
+              'import-export.step3.toggle',
+              'Import global settings (includes custom themes and compatibility fixes)'
+            )}
+            value={includeGlobal}
+            handleChange={() => setIncludeGlobal(!includeGlobal)}
+          />
+          <div className="ImportExportWizard__warningCard">
+            <WarningAmberIcon />
+            <div>
+              <strong>
+                {t(
+                  'import-export.step3.overwriteTitle',
+                  'Overwrite my current global settings'
+                )}
+              </strong>
+              <p>
+                {t(
+                  'import-export.step3.overwriteBody',
+                  'Off by default. Turn this on to replace your current Heroic settings with the ones in the backup.'
+                )}
+              </p>
+              <ToggleSwitch
+                htmlId="ie-overwrite-global"
+                title={t(
+                  'import-export.step3.overwrite',
+                  'Overwrite current global settings'
+                )}
+                value={overwriteGlobal}
+                handleChange={() => setOverwriteGlobal(!overwriteGlobal)}
+                disabled={!includeGlobal}
+              />
+            </div>
+          </div>
+          <p className="ImportExportWizard__meta">
+            {t('import-export.step3.counts', {
+              defaultValue: 'Includes {{fixes}}, {{themes}}.',
+              fixes: manifest.counts.fixesIncluded
+                ? t('import-export.step3.fixes-yes', 'compatibility fixes')
+                : t('import-export.step3.fixes-no', 'no compatibility fixes'),
+              themes: manifest.counts.themesIncluded
+                ? t('import-export.step3.themes-yes', 'custom themes')
+                : t('import-export.step3.themes-no', 'no custom themes')
+            })}
+          </p>
+        </>
+      )}
+    </section>
+  )
+}
